Changhao - added unit tests for GeneratePdfReport.jsx
